Remove AgentExecutor and refactor agent module structure
Remove executor.rs and reorganize agent functionality to eliminate complexity.
Convert max_iterations from usize to u32 for consistency across numeric types.
Comment out(for now) unused test code in tool.rs to simplify maintenance.

+++ src/agent/error.rs
@@ -14,7 +14,7 @@ pub enum AgentError {
     ResponseParsingError(String),
 
     #[error("Agent reached maximum iterations ({limit}) without completing task")]
-    MaxIterationsReached { limit: usize },
+    MaxIterationsReached { limit: u32 },
 
     #[error("Tool execution failed: {0}")]
     ToolError(#[from] Box<dyn std::error::Error + Send + Sync>),
